Specify exact BSON version dependency
Since the mongo and bson gems are versioned together, it should follow
that mongo depends on a specific version of bson. Otherwise, client
libraries that require a specific version of mongo, expecting the same
version of bson to be installed may get a different version. This is
particularly painful when a client library specifies a dependency on
both mongo and bson_ext using the same version number, which may not be
the latest, and the dependency from the mongo gem will actually install
a different version of bson than bson_ext.