Make it easier to use bson out of git with bundler
Requiring the entirety of the bson libraries in the gemspec can cause
them to establish state before all of the relevant pieces are in
place. In particular, loading bson can cause the bson gem to conclude
that bson_ext isn't available, when in fact its gemspec just hasn't
been parsed yet.