RUBY-583: add ability to run clusters with auth
This previously was failing as the ruby Process#spawn method was closing
the STDOUT file descriptor. Mongod as of (2.4.2) expects for STDIN and
STDOUT to be the first two file descriptors as the logging system
subsequently tries to reopen and reroute those hardcoded FD values.

You can read more about the issue here:

https://jira.mongodb.org/browse/SERVER-9362