RUBY-615 adding required notice file
Adding the required NOTICE file for compliance with our Apache 2.0 license. Since it's not required to include non-bundled dependencies, this file only contains our own copyright information.

With the changes I've made in RUBY-618, the only bundled dependency we have in this project now is with the BSON package from the Java driver. However, since it's also our own and licensed under Apache 2.0 we aren't required to list it here.