minor: relocating shared examples
i'm creating a new home for shared examples and updating the spec_helper to require up all shared examples dynamically. this will make it a lot easier to organize shared examples project wide later on and doesn't restrict us to keeping them all in the one file we had before.