Wait for replset optimes to catch up to the primary when testing
If we don't wait for the optimes to catch up and we do a force
stepdown, the old primary will just reelect itself, which doesn't make
for a great test of failover behavior.

This way, since the replset is caught up to the primary, there's no
need to pass force: true during the stepdown. We do need to bump the
timeout, though - 5s isn't necessarily enough for the nodes to
converge on a new primary.