minor: excluding pry-rescue for jruby
The latest release of JRuby (1.7.5) slipped out the door without a few binaries like 'jrubyc' for compiling java extensions. One of the dependencies for pry-rescue errantly uses this bin via a shell call to build its extension so as a result it doesn't install correctly under jruby-1.7.5.

The JRuby team has fixed this issue in their build system that caused jrubyc to be excluded, but they've yet to make another release. For now, I'm excluding this gem for java based ruby platforms.