Avoid sharing sockets between forked processes
Support for detecting when a socket was established in a different
process and automatically discarding it was added to this driver in
997d9b8ae2f5f391ef23dc4d3792e7d863675eb8, and refined in
a393557eeb8c1a9a049f4d289be51e50579ebf40.

Instances of the Node class maintain their own sockets for refreshing
replica set data; they need to be protected from reuse in the same way.