RUBY-986 Fallback to IPv4 if IPv6 isn't supported
Prior to this change, if the provided server address resolved to IPv6
but the mongo server wasn't configured for IPv6, the client would
continuously attempt to connect  to the IPv6 address until the
server_selection_timeout was hit and never fallback to IPv4.

This fixes the behavior to try all the different protocol families in
the list.

Fixes https://jira.mongodb.org/browse/RUBY-986.