RUBY-1332 Require that all servers in a topology satisfy wire protocol constraints.
This is required by the specs but wasn't done by the Ruby driver.
Instead the Ruby driver only checked wire protocol for the selected
server.

Example: Current min wire version is 2. If the primary is at version 2
and a secondary at version 0, and the application requests primary,
previously ruby driver will find the primary, check the version and
return that server and succeed/carry on. The spec says that secondary's
version 0 must mark the entire topology as unusable.

This commit doesn't mark the topology as unusable but it prevents a
server from being returned in such cases, passing the spec tests.

https://jira.mongodb.org/browse/RUBY-1355 was filed for follow-up work.