Always expose command_name as a string.
The driver itself documented command_name to be a string.
The specifications specify command_name be a string but they don't
know of Ruby hashes. However up until this point the Ruby driver
used both strings and symbols which is just silly.

This change is not a breaking one for clients that handled all
events, as they must have handled both strings and symbols.
It can be a breaking change for clients which looked for
specific commands, and specified those commands as symbols
(though there was no guarantee by the driver that the commands
wouldn't be given as strings by other code/at a different time.