Add options to disable built-in monitoring subscribers
Previously the user could only select between enabling and disabling
all monitoring for an instance of `Mongo::Client`.

That meant that if a custom subscriber was added through the global API

```ruby
Mongo::Monitoring::Global.subscribe(Mongo::Monitoring::COMMAND, MyCommandSubscriber.new)
```

and client monitoring was turned off, then the built-in subscribers
were not subscribed to the client's montoring, but neither was
the custom subscriber.

If client monitoring was turned on, then the custom subscriber was
subscribed to the client's monitoring, but so were most of the
buit-in subscribers with no option to turn them off.

Users who only want to see the debug output of their custom subscriber
may be annoyed by this and feel that it "clutters" their logs.

This commit adds the possibility to disable the built-in subscribers
one by one without breaking the gem's public API.