RUBY-3357 SSDLC Requirements (#2880)
* add more tasks to the release workflow

also, add a cleanup workflow for use while testing

* workflows don't have descriptions

* inputs are specified on the trigger, not the workflow

* skip the SBOM generation until we can figure it out

* use bson-ruby asset group, for testing

this is temporary until silk is able to grab the correct sbom lite from
the mongo-ruby-driver repo

* put code scan results in S3_ASSETS

and for testing, remove dry-run from s3 upload, to make sure it works

* restore the correct silk asset group

* refactor to try the new ruby actions in drivers-github-tools

* remove stray quotation mark

* use the canonical version of the ruby actions

* Indicate `dry_run` to be of type boolean

This is so it gets rendered as a checkbox. We'll default it to `true` to avoid accidental releases.